% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caribou.R
\docType{data}
\name{caribou}
\alias{caribou}
\title{GPS tracking data of a single caribou}
\format{An \code{ltraj} object representing the movement of a single caribou from GPS tracking, the data comprises 1793 fixes.}
\description{
GPS telemetry data for a single caribou tracked from 2000-01-10 to 2000-11-13 in northern British Columbia, Canada.The data were collected by the British Columbia Ministry of Environment.
}
\details{
The caribou data are stored as a single \code{'ltraj'} object and GPS fixes were attempted at a regular sampling frequency of 4 hours.
}
\examples{
data(caribou)
plot(caribou)
}
\keyword{datasets}

