% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.r
\docType{data}
\name{dem}
\alias{dem}
\title{Digital elevation data for northern BC, Canada}
\format{A \code{RasterLayer} with 1072 rows and 1102 columns.}
\description{
Digital elevation data, to be used alongside the caribou data for demonstrating field-based time geography.
}
\details{
A digital elevation model for Canada with a spatial resolution of 25 m. Full details of this data can be obtained from the GeoBC website. It is projected using the NAD83 UTM Zone 9N projection.
}
\keyword{datasets}

