% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dynppa}
\alias{dynppa}
\title{Dynamic PPA Measure of Animal Space Use}
\usage{
dynppa(traj, tol = max(ld(traj)$dt, na.rm = TRUE), dissolve = TRUE,
  proj4string = CRS(as.character(NA)), ePoints = 360, ...)
}
\arguments{
\item{traj}{an object of the class \code{ltraj} which contains the time-stamped
movement fixes of the first object. Note this object must be a \code{type II
ltraj} object. For more information on objects of this type see \code{
help(ltraj)}.}

\item{tol}{parameter used to filter out those segments where the time between fixes is overly
large (often due to irregular sampling or missing fixes); which leads to an overestimation of the
activity space via the PPA method. Default is the maximum sampling interval from \code{traj}.}

\item{dissolve}{(logical) whether or not to dissolve output elliplse polygons to create a single
output polygon, or keep the individual segment PPA ellipses. Default = TRUE.}

\item{proj4string}{a string object containing the projection information to be passed included in the output
\code{SpatialPolygonsDataFrame} object. For more information see the \code{CRS-class} in the packages
\code{sp} and \code{rgdal}. Default is \code{NA}.}

\item{ePoints}{number of vertices used to construct each PPA ellipse. More points will necessarily provide
a more detailed ellipse shape, but will slow computation; default is 360.}

\item{...}{additional parameters to be passed to the function \code{dynvmax}. For example, should include
   \code{method} and/or \code{dynamic} parameters, see the documentation for \code{dynvmax} for more detailed
   information on what to include here.}
}
\value{
This function returns a \code{SpatialPolygonsDataFrame} representing the dynamic PPA measure of the accessibility
  space of an individual animal.
}
\description{
The function \code{dynppa} computes the (dynamic) PPA measure of the accessibility space of an animal.
  The PPA method can be thought as an alternative view on the home range; one that explicitly considers the
  spatial and temporal constraints on movement given known telemetry fixes, and a (dynamic) measure of maximum
  mobility - termed Vmax. The PPA method incorporates dynamic behaviour into the calculation of the vmax parameter
  used to delineate the original version of the PPA method, but the original method is still an option here.
}
\details{
The function \code{dyn.ppa} represents an extension to an existing PPA method (Long and Nelson, 2012).
  Dynamic calculation of the PPA method improves upon the original version by flexibly modelling the vmax
  parameter according to wildlife behaviour. See the function \code{dyn.vmax} for more information on how
  to incorporate dynamic behaviour into the vmax parameter estimation.
}
\references{
Long, JA, Nelson, TA. (2012) Time geography and wildlife home range delineation. \emph{Journal of Wildlife
  Management}, 76(2):407-413.\cr \cr
  Long, JA, Nelson, TA. (2014) Home range and habitat analysis using dynamic time geography. \emph{Journal of
  Wildlife Management}. Accepted: 2014-12-03.\cr
}
\seealso{
dynvmax
}

