% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynvmax.R
\name{dynvmax}
\alias{dynvmax}
\title{Dynamic Calculation of the Vmax Parameter}
\usage{
dynvmax(traj, dynamic = "NA", w = 9, class.col = "dt",
  method = "Robson", k = 5, alpha = 0.05, manualVmax = NA,
  vmaxtrunc = NA)
}
\arguments{
\item{traj}{an object of the class \code{ltraj} which contains the time-stamped
movement fixes of the first object. Note this object must be a \code{type II
ltraj} object. For more information on objects of this type see \code{
help(ltraj)}.}

\item{dynamic}{one of \code{'NA'}, \code{'focal'}, \code{'cumulative'}, or \code{'class'}; which signifies 
whether or how to dynamically compute the Vmax parameter. See \bold{Details} for more information on 
each of the choices.}

\item{w}{(optional) window size (only used with \code{dynamic = 'focal'} or \code{'cumulative'}).}

\item{class.col}{(optional) character indicating the name of the column in the \code{infolocs} dataframe
of \code{traj} containing the categorized behavioural states of the animal (which can be stored as
a character or numeric column).}

\item{method}{method for computing the Vmax parameter dynamically; can be one of several options:
-- \code{"Robson"} for the Robson & Whitlock (1964) method,\cr
-- \code{"RobsonLL"} for the R & W (1964) lower \eqn{(1-\alpha)*100\%} C.I. limit, \cr
-- \code{"RobsonUL"} for the R & W (1964) upper \eqn{(1-\alpha)*100\%} C.I. limit, \cr
-- \code{"vanderWatt"} for the van der Watt (1980) method, \cr
-- \code{"vanderWattLL"} for the van der Watt (1980) lower \eqn{(1-\alpha)*100\%} C.I. limit, \cr
-- \code{"vanderWattUL"} for the van der Watt (1980) upper \eqn{(1-\alpha)*100\%} C.I. limit. \cr}

\item{k}{(optional) value for the \emph{k} parameter in the van der watt (1980) method; default is 5.}

\item{alpha}{(optional) value for the \eqn{\alpha} parameter if using upper or lower C.I. methods; default is 0.05.}

\item{manualVmax}{(optional) Character name of column in \code{traj} storing user input column of vmax values (typically call the column dynVmax).}

\item{vmaxtrunc}{(optional) due to irregular sampling intervals, or errors in GPS location, or other
effects, the calculation of the vmax parameter through the statistical methods outlined above can be
heavily influenced by high outliers. Thus, it may be useful to exclude those segments from calculation 
of the dynamic Vmax parameter. Default is \code{NA}.}
}
\value{
This function returns the original \code{traj} object with a new column -- \code{dynVmax} in the \code{infolocs} dataframe
  containing the dynamic vmax parameter for each trajectory segment.
}
\description{
The function \code{dynvmax} computes a dynamic version of the Vmax parameter for the PPA method. It can be used to incorporate changes in animal movement behaviour into the PPA method caluculation to better model that area accessible to an individual animal given the set of known telemetry locations in space and time.
}
\details{
The function \code{dynvmax} represents an intermediary function used to extend and improve upon an existing PPA home range method (Long and Nelson, 2012) as described in the paper (Long and Nelson, 2014). Four options are available for computing the vmax parameter dynamically and are passed into the \code{dynvmax} function using \code{dynamic} option.\cr 
 \cr 1) \code{NA} -- if \code{dynamic} = \code{'NA'} (the default) the function estimates the original, 
         non-dynamic estimate of Vmax which is a global estimate, as per Long & Nelson (2012). 
 \cr 2) \code{focal} -- a moving window approach whereby a window of size \code{w} is moved along the 
       trajectory and vmax computed dynamically within each window and assigned to the central segment.
 \cr 3) \code{cumulative} -- A moving window of size \code{w} is again used, only in this case the value 
       is assigned to the end segment. This represents the vmax calculation of the previous \code{w} segments.
 \cr 4) \code{class} -- A priori analysis (e.g., obtained via state-space models, or from expert knowledge) 
       is used to identify discrete behavioural states in the telemetry data and these stored in a column 
       which is then passed into the function.\cr\cr 
The \code{class} method is the preferred choice, as it allows the use of more sophisticated models for identifying behavioural shifts in telemetry data where we would expect to see clear differences in the Vmax parameter based on changing movement behaviour.\cr\cr 
The use of the \code{'focal'} or \code{'cumulative'} dynamic methods uses a moving window approach, which is sensitive to edge effects at the initial and ending times of the trajectory. Thus, the dynamic Vmax parameter is only computed for those segments that have a valid window and the dataset is shrunk by \code{w-1} segments.
}
\examples{
m3R <- dynvmax(m3,dynamic='focal',method='Robson')
m3V <- dynvmax(m3,dynamic='focal',method='vanderWatt')
m3c <- dynvmax(m3,dynamic='cumulative')

}
\references{
Long, JA, Nelson, TA. (2012) Time geography and wildlife home range delineation. \emph{Journal 
  of Wildlife Management}. 76(2):407-413.\cr\cr
  Long, JA, Nelson, TA. (2015) Home range and habitat analysis using dynamic time geography. \emph{Journal of -
  Wildlife Management}. 79(3):481-490.\cr\cr
  Robson, DS, Whitlock, JH. (1964) Estimation of a truncation point. \emph{Biometrika}
  51:33-39.\cr\cr
  van der Watt, P. (1980) A note on estimation bounds of random variables. \emph{Biometrika}
  67(3):712-714.\cr
}
\seealso{
dynppa
}
