% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estc2.R
\name{estc2}
\alias{estc2}
\title{Numerically estimate c2 parameter}
\usage{
estc2(traj, tl, timefun = "exp", sigma = 0, min = 0, max = 1,
  rand = NA, niter = 10, tolerance = 0.01, plot = TRUE)
}
\arguments{
\item{traj}{animal movement trajectory in the form of an \code{ltraj} object, see package \code{adehabitatLT}}

\item{tl}{a \code{TransitionLayer} object}

\item{timefun}{method for converting time into probability (see \code{fbtgUD}); one of:\cr 
-- \code{'inverse'} \eqn{= \frac{1}{c_2 * t}},\cr
-- \code{'inverse2'} \eqn{= \frac{1}{(c_2 * t)^2}},\cr
-- \code{'exp'} \eqn{= \exp{-c_2 * t}},\cr
-- \code{'norm'} \eqn{= \exp{-c_2 * t^2}},\cr
-- \code{'rootexp'} \eqn{= \exp{-c_2 * \sqrt{t}}},\cr
-- \code{'pareto'} \eqn{= \exp{-c_2 * \log{t}}},\cr
-- \code{'lognorm'}\eqn{= \exp{-c_2 * \log{t^2}}}.\cr}

\item{sigma}{location uncertainty parameter (see fbtgUD)}

\item{min}{lower bound for \code{c2} testing range (default = 0)}

\item{rand}{if \code{NA} (the default) every second segment is evauluated (n/2), otherwise an integer indicating how many random segments to test.}

\item{niter}{used to define maximum number of iterations of golden-search routine}

\item{tolerance}{used to define precision of golden search routine (i.e., routine stops when the absolute difference between two consecutive test points is below this value).}

\item{plot}{logical, whether or not to plot the log-likelihood curve.}

\item{min}{upper bound for \code{c2} testing range (default = 1)}
}
\value{
This function returns a numerical value estimate for \code{c2} associated with the maximum of the log-likelihood.
}
\description{
Numerically estimate the \code{c2} parameter for \code{fbtgUD} using a data-driven approach.
}
\details{
THe estimation of the \code{c2} parameter for \code{fbtgUD} model takes an identical approach to that used with Brownian bridges, as originally proposed by Horne et al. (2007). A leave-one-out estimation technique is used, which essentially removes fixes and then estimates the fbtgUD surface between the two adjacent fixes  -- termed a segment -- and computes the probability associated with the missing fix. The \code{c2} value is returned that numerically maximizes the log-likelihood (for a given \code{timefun} -- see details in \code{fbtgUD}) for \code{length.out} evenly spaced values of \code{c2} between (user-defined) \code{min} and \code{max}. With \code{plot = TRUE} the user can verify that the chosen range of potential \code{c2} values is appropriate, and if not, retry using a differnt range. The process is computationally demanding and is highly dependent on the number of 'segments' used. The parameter \code{rand} can be used to adjust the number of segments used to minimize computational time. If \code{rand = NA} it removes every second fix, and estimates \code{c2} based on these n/2 segments (the default). Otherwise \code{rand} can be passed in as an integer, and \code{rand} randomly selected segments will be chosen to estimate \code{c2}. This is beneficial for trajectories with many fixes, where it might be useful to choose \code{rand <<< n/2}. Parrallelization is possible to further decrease computational time. This can be implemented by choosing an appropriate integer value for the \code{parallel} parameter (implemented using the \code{foreach} package).
}
\seealso{
fbtgUD
}
