% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbtgTS.R
\name{fbtgTS}
\alias{fbtgTS}
\title{Field-based Time Geography Utilization Distribution}
\usage{
fbtgTS(traj, tk, surf, sigma, timefun = "inverse", c2 = 1, clipPPS = TRUE)
}
\arguments{
\item{tk}{time to compute the probability surface, \code{ta < tk < tb}.}

\item{timefun}{method for converting movement time into probability; one of: \code{'inverse' (default),'inverse2','exp','norm','rootexp','pareto','lognorm'}, see details.}

\item{c2}{Parameter input into \code{timefun} default=1, see details.}

\item{A}{coordinates of origin point of a movement segment in the form \code{c(x,y,t)}.}

\item{B}{coordinates of end point of a movement segment in the form \code{c(x,y,t)}.}

\item{k}{number of time slices between fixes upon which to estimate the UD, default=10.}
}
\value{
This function returns a \code{RasterLayer} which can be used to estimate the UD of an animal.
}
\description{
Calculate the utilization distribution of an animal using the field-based time geographic model.
}
\details{
Calculates the field-based time geography probability surface for an animal for a chosen time - often called a time slice. Field-based time geography is based on an underlying resistance surface, which constratins potential movement by the animal between two location fixes. The output probability surface represents the probability of finding an animal in a given location at the input time, based on the field-based time geography movement model.
}
\seealso{
ppa, fbtgTS, volras
}

