% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalPfun.R
\name{internalPfun}
\alias{internalPfun}
\title{Functions for converting time to probabilities}
\usage{
internalPfun(x, timefun, c2)
}
\arguments{
\item{x}{a numeric object (i.e., a \code{RasterLayer}) upon which to compute the probabilities.}

\item{timefun}{method for convertingtime into probability; one of:\cr 
-- \code{'inverse'} \eqn{= \frac{1}{c_2 * t}},\cr
-- \code{'inverse2'} \eqn{= \frac{1}{(c_2 * t)^2}},\cr
-- \code{'exp'} \eqn{= \exp{-c_2 * t}},\cr
-- \code{'norm'} \eqn{= \exp{-c_2 * t^2}},\cr
-- \code{'rootexp'} \eqn{= \exp{-c_2 * \sqrt{t}}},\cr
-- \code{'pareto'} \eqn{= \exp{-c_2 * \log{t}}},\cr
-- \code{'lognorm'}\eqn{= \exp{-c_2 * \log{t^2}}}.\cr}

\item{c2}{Parameter input into \code{timefun} default=1.}
}
\value{
This function returns an object identical to \code{x} with \code{timefun} applied to its values.
}
\description{
Internal function for computing movement probabilities.
}
\details{
Used in the fbtg function.
}
\keyword{internal}

