% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalSEG.R
\name{internalSEG}
\alias{internalSEG}
\title{Segment Function for fbtg}
\usage{
internalSEG(j, df, tl, k, sigma, timefun, c2, clipPPS)
}
\arguments{
\item{j}{index of the segment.}

\item{df}{trajectory dataframe (from \code{ld(traj)})}

\item{tl}{\code{TransitionLayer} modelling movement costs.}

\item{k}{number of steps for use in \code{internalTS}.}

\item{sigma}{parameter for modelling location uncertainty (s.d. of Guassian location error).}

\item{timefun}{function for modelling probabilities from time, input into Pfun.}

\item{c2}{coefficient for modelling probabilities from time, input into Pfun.}

\item{clipPPS}{whether or not to clip to the PPS (default = TRUE).}
}
\value{
This function returns a \code{RasterLayer} of the probabilities for time slice ta.
}
\description{
Internal function for computing Pi for segments.
}
\details{
Used in the \code{fbtg} function.
}
\keyword{internal}
