% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccCost.R
\name{JaccCost}
\alias{JaccCost}
\title{Accumulated cost function - Modified by Jed}
\usage{
JaccCost(x, fromCoords, mode)
}
\arguments{
\item{x}{a \code{TransitionLayer} object}

\item{fromCoords}{location coordinates for accumulated cost calculation}

\item{mode}{one of \code{'in'} or \code{'out'} signifying if the direction of accumulation is towards or away from the location.}
}
\value{
This function returns a \code{RasterLayer} which has the accumulated costs.
}
\description{
Modified Accumulated cost function.
}
\details{
Modified Accumulated Cost function from the gdistance accCost function.
}
\keyword{internal}

