% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jppa}
\alias{jppa}
\title{Joint Potential Path Area  of Two Animals}
\usage{
jppa(traj1, traj2, t.int = 0.1 *
  as.numeric(names(sort(-table(ld(traj1)$dt)))[1]), tol = max(ld(traj1)$dt,
  na.rm = T), dissolve = TRUE, proj4string = CRS(as.character(NA)),
  ePoints = 360, ...)
}
\arguments{
\item{traj1}{an object of the class \code{ltraj} which contains the time-stamped
movement fixes of the first object. Note this object must be a \code{type II
ltraj} object. For more information on objects of this type see \code{
help(ltraj)}.}

\item{traj2}{same as \code{traj1}.}

\item{t.int}{(optional) time parameter (in seconds) used to determine the frequency of time slices
used to delineate the joint activity space. Default is 1/10th of the mode of the temporal sampling
interval from \code{traj1}. Smaller values for \code{t.int} will result in smoother output polygons.}

\item{tol}{(optional) parameter used to filter out those segments where the time between fixes is overly
large (often due to irregular sampling or missing fixes); which leads to an overestimation of the
activity space via the PPA method. Default is the maximum sampling interval from \code{traj1}.}

\item{dissolve}{logical parameter indicating whether (\code{=TRUE}; the default) or not (\code{=FALSE})
to return a spatially dissolved polygon of the joint activity space.}

\item{ePoints}{number of vertices used to construct each PPA ellipse. More points will necessarily provide
a more detailed ellipse shape, but will slow computation; default is 360.}

\item{proj4string}{a string object containing the projection information to be passed included in the output
\code{SpatialPolygonsDataFrame} object. For more information see the \code{CRS-class} in the packages
\code{sp} and \code{rgdal}. Default is \code{NA}.}

\item{...}{additional parameters to be passed to the function \code{dynvmax}. For example, should include options for
   \code{dynamic} and \code{method}; see the documentation for \code{dynvmax} for more detailed information
   on what to include here.}
}
\value{
This function returns a \code{SpatialPolygonsDataFrame} representing the joint accessibility space between
  the two animals.
}
\description{
The function \code{jppa} computes the joint accessibility space between two animals. It can be used
  to map (as a spatial polygon) the area that could have been jointly accessed by two individual animals
  in space ant time. The jPPA represents a spatial measure of spatial-temporal interaction.
}
\details{
The function \code{jppa} can be used to map areas of potential interaction between two animals.
  Specifically, this represents a measure of spatial overlap that also considers the temporal sequencing
  of telemetry points. In this respect it improves significantly over static measures of home range overlap,
  often used to measure static interaction, and can be considered as a spatial measure of dynamic interaction.
}
\seealso{
dynvmax, dynppa
}

