% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3.R
\docType{data}
\name{m3}
\alias{m3}
\title{GPS tracking data of a single caribou}
\format{An \code{ltraj} object representing the movement of a single caribou from GPS tracking, the data comprises 327 fixes.}
\description{
GPS telemetry data for a single caribou tracked during the summer season (July-August 2000) in northern British Columbia, Canada.The data were collected by the British Columbia Ministry of Environment.
}
\details{
The caribou data are stored as a single \code{'ltraj'} object and GPS fixes were attempted at a regular sampling frequency of 4 hours.
}
\examples{
data(m3)
plot(m3)
}
\keyword{datasets}

