% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppaEllipse.r
\name{ppaEllipse}
\alias{ppaEllipse}
\title{PPA Ellipse}
\usage{
ppaEllipse(x, y, a, b, theta, steps)
}
\arguments{
\item{x}{first coordinate}

\item{y}{second coordinate}

\item{a}{semi-major axis}

\item{b}{semi-minor axis}

\item{theta}{rotation angle of the ellipse (in radians)}

\item{steps}{number of segments, from ePoints parameter in \code{dyn.ppa.hr}}
}
\value{
This function returns a polygon ellipse.
}
\description{
Internal ellipse calculation function.
}
\details{
Internal function for calculating ellipses in time geographic analysis.
}
\seealso{
dynppa
}
\keyword{internal}

