% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppv.R
\name{ppv}
\alias{ppv}
\title{Potential path volume: 3D home range}
\usage{
ppv(traj, zcol = "z", vmax, vox)
}
\arguments{
\item{traj}{an object of the class \code{ltraj} which contains the time-stamped movement fixes of the first object. Note this object must be a \code{type II ltraj} object. It also requires a column in the \code{infolocs} which provides the 3rd Dimension (height) infromation. For more information on objects of this type see \code{help(ltraj)}.}

\item{zcol}{(character string) giving the name of the column in the \code{infolocs} where the height attribute data is stored.}

\item{vmax}{either a single vmax value giving the vmax value, or a vector, length n-1, giving the vmax value for each segment.}

\item{vox}{size of voxels in map units (e.g., if {x,y,z} are in meters then vox should be in meters).}

\item{tol}{parameter used to filter out those segments where the time between fixes is overly large (often due to irregular sampling or missing fixes); which leads to an overestimation of the activity space via the PPV method. Default is the maximum sampling interval from \code{traj}.}
}
\value{
This function returns a XX representing the ppv 3D home range measure, which can be interpreted as the 3D accessibility space of an individual animal.
}
\description{
The function \code{ppv} computes the potential path volume (PPV) 3D home range. It is also a measure of the accessibility space of an animal in 3D. The PPV method considers the spatial and temporal constraints on movement given known telemetry fixes, and a (dynamic) measure of maximum mobility - termed Vmax.
}
\details{
The function \code{ppv} represents an extension to an existing 2D home range method (Long and Nelson, 2012) for 3D. ADD MORE HERE.
}
\references{
Demsar, U, Long, JA (in review) Potential Path Volume (PPV): a geometric estimator for space use in 3D. In review at Movement Ecology. 
  
  Long, JA, Nelson, TA. (2012) Time geography and wildlife home range delineation. \emph{Journal of Wildlife
  Management}, 76(2):407-413.\cr \cr
}
\seealso{
dynvmax dynppa
}
