% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspUD.R
\name{rspUD}
\alias{rspUD}
\title{Randomised Shortest Path Utilization Distribution}
\usage{
rspUD(traj, r, theta, theta.col)
}
\arguments{
\item{traj}{animal movement trajectory in the form of an \code{ltraj} object, see package \code{adehabitatLT}}

\item{r}{a \code{RasterLayer} object describing the preference/affinity of the landscape. May be the result of a resource selection function, or other analyses. Note: Higher values should be associated higher affinity.}

\item{theta}{a global value for theta.}

\item{theta.col}{character string of the name of the column containing values of theta for each segment (only used if theta is not provided).}
}
\value{
This function returns a \code{RasterLayer} which can be used to estimate the UD of an animal.
}
\description{
Calculate the utilization distribution of an animal based on randomised shortest paths between fixes.
}
\details{
Still under development, but should work.
}
\seealso{
ppa, fbtgUD
}
