% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segtraj.R
\name{segtraj}
\alias{segtraj}
\title{Simple trajectory segmentation}
\usage{
segtraj(traj, k = 2, style = "fisher", col = "dist")
}
\arguments{
\item{traj}{an object of the class \code{ltraj} which contains the time-stamped movement fixes of the first object. Note this object must be a \code{type II ltraj} object. For more information on objects of this type see \code{help(ltraj)},}

\item{k}{(integer) the number of output classes desired from the segmentation (default = 2),}

\item{style}{(character)the algorithm for identifying class breaks (default = 'fisher'),}

\item{col}{the name (as a character string) of the column upon which to perform the segmentation (default = 'dist')}
}
\value{
An \code{ltraj} object with a new infolocs column 'class'.
}
\description{
Using the \code{classIntervals} function from the \code{classInt} package perform trajectory segmentation based on a variety of simple algorithms.
}
\details{
The segementation is based on a single trajectory attribute (default is the movement distance). The default method for performing the segmentation is the Fisher-Jenks algorithm, but many are available (see the full list in the documentation for the \code{classIntervals} function).
}
