% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma.r
\name{sma}
\alias{sma}
\title{Slow movement areas}
\usage{
sma(traj, sma.keep = 1, sma.tol = 1, tol = max(ld(traj)$dt, na.rm = TRUE),
  proj4string = CRS(as.character(NA)), ePoints = 360, ...)
}
\arguments{
\item{traj}{an object of the class \code{ltraj} which contains the time-stamped
movement fixes of the first object. Note this object must be a \code{type II
ltraj} object. For more information on objects of this type see \code{help(ltraj)}.}

\item{sma.keep}{an integer value indicating the number of slow movement areas to delineate, default is 1.}

\item{sma.tol}{a value <= 1 indicating used when sma.keep > 1 to define how much overlap is allowed between SMA's, if sma.tol=1 no overlap is allowed, if sma.tol=0, any and all overlap is allowed. Typically something in between is most useful. Defaults to 1.}

\item{tol}{(optional) parameter used to filter out those segments where the time between fixes is overly 
large (often due to missing fixes); which leads to an overestimation of the 
activity space via the PPA method. Default is the maximum sampling interval from \code{traj1}.}

\item{proj4string}{a string object containing the projection information to be passed included in the output 
\code{SpatialPolygonsDataFrame} object. For more information see the \code{CRS-class} in the packages
\code{sp} and \code{rgdal}. Default is \code{NA}.}

\item{ePoints}{number of vertices used to construct each PPA ellipse. More points will necessarily provide
a more detailed ellipse shape, but will slow computation; default is 360.}

\item{...}{additional parameters to be passed to the function \code{dynvmax}. For example, should include options for
\code{dynamic} and \code{method}; see the documentation for \code{dynvmax} for more detailed information 
on what to include here.}
}
\value{
This function returns a \code{SpatialPolygonsDataFrame} representing the joint accessibility space between
  the two animals.
}
\description{
The function \code{sma} computes the areas representing slow movement areas as described in the paper Nelson et al. (2014). Slow movement areas represent areas of sustained or intense habitat use, related to slow movement behaviours. Slow movement areas are defined by counting consecutive fixes within time geographic ellipses, and represented spatially as spatial polygons that are the union of included telemetry fixes within the slow movement area.
}
\details{
The function \code{sma} can be used to map slow movement areas identifiable from wildlife telemetry data.of potential interaction between two animals. Slow movement areas can be ranked, according to their importance, which equates to consecutive time spent in an area. That is, the first slow movement area will be the area where the animal stayed the longest, and so on. Thus, slow movement areas can be useful for identifying where encamped behaviour or intensively exploited habitat on the landscape.
}
\references{
Nelson, T.A., Long, J.A., Laberee, K., Stewart, B.P. (2015) A time geographic approach for delineating areas of sustained wildlife use. Annals of GIS. 21(1): 81-90.
}
\seealso{
dynvmax, dynppa
}
