% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgkde.r
\name{tgkde}
\alias{tgkde}
\title{Time geographic kernel density estimation}
\usage{
tgkde(traj, disfun = "inv", c2 = 1, grid = NA, ...)
}
\arguments{
\item{traj}{an object of the class \code{ltraj} which contains the time-stamped movement fixes of the first object. Note this object must be a \code{type II ltraj} object. For more information on objects of this type see \code{ help(ltraj)}.}

\item{disfun}{one of { \code{'inv', 'inv2', 'exp', 'norm'} } representing the shape of the distance decay function used to model movement probabilities within the geoellipse (see Downs et al. (2011) for more information).}

\item{c2}{parameter used to control shape of different distance decay functions, default = 1.}

\item{grid}{spatial resolution (pixel size) of output utilization raster in appropriate units. Default is chosen based on the x and y range of the input telemetry data. Alternatively, a \code{RasterLayer} can be passed in upon which the UD is computed.}

\item{...}{additional parameters to be passed to the function \code{dynvmax}. For example, should include options for \code{dynamic} and \code{method}; see the documentation for \code{dynvmax} for more detailed information on what to include here.}
}
\value{
This function returns a \code{RasterLayer} representing the utilization distribution of the animal
}
\description{
Compute the time geographic kernel density estimate of an animals utilization distribution following the methods described in Downs et al. (2011).
}
\details{
The function \code{tgkde} can be used to delineate an animals home range using the time geographic kernel density estimation procedure described by Downs et al. (2011). Specifically, it modifies the shape of the traditional kernel to consider the time geographic limits of movement opportunity - termed the geoellipse by Downs, which is analagous to the potential path area concept described by Long & Nelson (2012,2015). Several basic functions -- including inverse distance, inverse distance squared, exponential, and normal -- can be used to quantify movement probabilities within the geoellipse. The output is then the utilization distribution of the animal, confined to the accessibility space defined by the potential path area home range as in Long & Nelson (2012,2015).
  The function \code{volras} can be used to extract volume contours (e.g., 95% volume contour) which are commonly used to delineate home ranges based on the utilization distribution.
}
\references{
Downs, J.A., Horner, M.W., Tucker, A.D. (2011) Time-geographic density estimation for home range analysis. Annals of GIS. 17(3): 163-171.
}
\seealso{
dynvmax, dynppa, volras
}
