% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caribou.R, R/dem.r
\docType{data}
\name{caribou}
\alias{caribou}
\title{GPS tracking data of a single caribou}
\format{An \code{ltraj} object representing the movement of a single caribou from GPS tracking, the data comprises 1793 fixes.}
\description{
GPS telemetry data for a single caribou tracked from 2000-01-10 to 2000-11-13 in northern British Columbia, Canada.The data were collected by the British Columbia Ministry of Environment.

Digital elevation data, to be used alongside the caribou data for demonstrating field-based time geography.
}
\details{
The caribou data are stored as a single \code{'ltraj'} object and GPS fixes were attempted at a regular sampling frequency of 4 hours.

A digital elevation model for Canada with a spatial resolution of 25 m. Full details of this data can be obtained from the GeoBC website. It is projected using the NAD83 UTM Zone 9N projection.
}
\examples{
data(caribou)
plot(caribou)
data(dem)
plot(dem)
}
\keyword{datasets}

