% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.r
\docType{data}
\name{dem}
\alias{dem}
\title{Digital elevation data for study area in BC, Canada}
\format{A \code{RasterLayer} with 275 rows and 347 columns.}
\description{
Digital elevation data, to be used alongside the caribou data for demonstrating field-based time geography.
}
\details{
A digital elevation model for the Caribou study area at 100 m. It has been clipped to the area where telemetry data is present. It is projected using the NAD83 UTM Zone 8N projection.
}
\examples{
data(dem)
plot(dem)
}
\keyword{datasets}
