% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esttheta.R
\name{esttheta}
\alias{esttheta}
\title{Numerically estimate theta parameter for rspUD}
\usage{
esttheta(traj, r, min = 0, max = 1, niter = 100, tolerance = 0,
  rand = 10, dmin = res(r)[1], plot = TRUE)
}
\arguments{
\item{traj}{animal movement trajectory in the form of an \code{ltraj} object, see package \code{adehabitatLT}}

\item{r}{a \code{RasterLayer} object describing the preference/permeability of the landscape. May be the result of a resource selection function, or other analyses. Note: Higher values should be associated higher preference or permeability.}

\item{min}{lower bound for \code{theta} testing range (default = 0)}

\item{max}{upper bound for \code{theta} testing range (default = 1)}

\item{niter}{used to define maximum number of iterations of golden-search routine.}

\item{tolerance}{used to define precision of golden search routine (i.e., routine stops when the absolute difference between two consecutive test points is below this value).}

\item{rand}{an integer indicating how many random segments to test.}

\item{dmin}{(defaults to \code{res(r)[1]}) the minimum movement distance to use as random samples. See details.}

\item{plot}{logical, whether or not to plot the log-likelihood curve.}
}
\value{
This function returns an estimate for the theta parameter which can be used with the rspUD function.
}
\description{
Estimate theta parameter using leave-one-out estimation procedure for randomized shortest path utilization distributions.
}
\details{
Still under development, currently it is WAY TO SLOW, looking at Panzachi and Saerens to see how to speed up (i.e., by not using passage function).
}
\seealso{
rspUD
}
