% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbtgTS.R
\name{fbtgTS}
\alias{fbtgTS}
\title{Field-based Time Geography Utilization Distribution}
\usage{
fbtgTS(traj, tk, tl, sigma, timefun = "inverse", c2 = 1, clipPPS = TRUE)
}
\arguments{
\item{traj}{animal movement trajectory in the form of an \code{ltraj} object, see package \code{adehabitatLT}}

\item{tk}{time to compute the probability surface, \code{ta < tk < tb}.}

\item{tl}{a \code{TransitionLayer} object}

\item{timefun}{method for convertingtime into probability; one of:\cr 
-- \code{'inverse'} \eqn{= \frac{1}{c_2 * t}},\cr
-- \code{'inverse2'} \eqn{= \frac{1}{(c_2 * t)^2}},\cr
-- \code{'exp'} \eqn{= \exp{-c_2 * t}},\cr
-- \code{'norm'} \eqn{= \exp{-c_2 * t^2}},\cr
-- \code{'rootexp'} \eqn{= \exp{-c_2 * \sqrt{t}}},\cr
-- \code{'pareto'} \eqn{= \exp{-c_2 * \log{t}}},\cr
-- \code{'lognorm'}\eqn{= \exp{-c_2 * \log{t^2}}}.\cr}

\item{c2}{Parameter input into \code{timefun} default=1, see details.}

\item{k}{number of time slices between fixes upon which to estimate the UD, default=10.}
}
\value{
This function returns a \code{RasterLayer} which can be used to estimate the UD of an animal.
}
\description{
Calculate the utilization distribution of an animal using the field-based time geographic model.
}
\details{
Calculates the field-based time geography probability surface for an animal for a chosen time - often called a time slice. Field-based time geography is based on an underlying resistance surface, which constratins potential movement by the animal between two location fixes. The output probability surface represents the probability of finding an animal in a given location at the input time, based on the field-based time geography movement model.
}
\seealso{
ppa, fbtgTS, volras
}

