% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalTS.R
\name{internalTS}
\alias{internalTS}
\title{Time Slice Function for fbtg}
\usage{
internalTS(ta, t, Tai, Tib, tl, Tshort, sigma, timefun, c2, clipPPS)
}
\arguments{
\item{ta}{time point of the slice.}

\item{t}{total duration of the segment.}

\item{Tai}{accumulated costs from anchor A to location i.}

\item{Tib}{accumulated costs from location i to anchor B.}

\item{tl}{\code{TransitionLayer} object used to compute Tai, Tib}

\item{sigma}{parameter for modelling location uncertainty (s.d. of Guassian location error).}

\item{timefun}{function for modelling probabilities from time, input into Pfun.}

\item{c2}{coefficient for modelling probabilities from time, input into Pfun.}

\item{clipPPS}{whether or not to clip to the PPS (default = TRUE).}
}
\value{
This function returns a \code{RasterLayer} of the probabilities for time slice ta.
}
\description{
Internal function for computing time slices.
}
\details{
Used in the \code{fbtg} function.
}
\keyword{internal}
