% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volras.r
\name{volras}
\alias{volras}
\title{Volume contour from Raster}
\usage{
volras(x, percent = 95, maxpixels = 10000)
}
\arguments{
\item{x}{a \code{RasterLayer}}

\item{percent}{a percent value to get the volume contour, e.g., 95.}

\item{simplify}{(logical) whether or not to simplify the output contour (uses \code{gSimplify}).}
}
\value{
A \code{SpatialPolygonsDataFrame}.
}
\description{
Compute a percent volume contour from a raster UD.
}
\details{
The volras function simply modifies the getvolumeUD function from the package \code{adehabitatHR} developed by C. Calenge. Basically, I was getting errors when using the getvolumeUD function, so I modified that function, and added the simplify polygon functionality.
}

